Refactored controler to controller
