+++ controller/checkRedis.js
